+++ src/refrigerator/convertibleDrawer.ts
@@ -38,7 +38,7 @@ export class ConvertibleDrawer {
       clientId:       platform.config.clientId,
       clientSecret:   platform.config.clientSecret,
       redirectUri:    platform.config.redirectUri,
-      debug:          platform.config.debugLogging || false,
+      debug:          platform.config.debug || false,
     });
 
     //=====================================================================================