+++ src/refrigerator/freezer.ts
@@ -34,7 +34,7 @@ export class Freezer {
       clientId:       platform.config.clientId,
       clientSecret:   platform.config.clientSecret,
       redirectUri:    platform.config.redirectUri,
-      debug:          platform.config.debugLogging || false,
+      debug:          platform.config.debug || false,
     });
 
     this.client.debug('Adding Freezer Thermostat');