+++ src/refrigerator/iceMaker.ts
@@ -30,7 +30,7 @@ export class IceMaker {
       clientId:       platform.config.clientId,
       clientSecret:   platform.config.clientSecret,
       redirectUri:    platform.config.redirectUri,
-      debug:          platform.config.debugLogging || false,
+      debug:          platform.config.debug || false,
     });
 
     this.client.debug('Adding Ice Maker Switch');