+++ src/refrigerator/turboCoolMode.ts
@@ -33,7 +33,7 @@ export class TurboCoolMode {
       clientId:       platform.config.clientId,
       clientSecret:   platform.config.clientSecret,
       redirectUri:    platform.config.redirectUri,
-      debug:          platform.config.debugLogging || false,
+      debug:          platform.config.debug || false,
     });
 
     //=====================================================================================